Added Check Out Razor View

+++ Hotel-Domain/Entities/Orders/Order.cs
@@ -25,7 +25,7 @@ namespace Hotel_Domain.Entities.Orders
         public long OrderSum { get; set; }
 
         [Required(ErrorMessage = "لطفا {0} را وارد کنید")]
-        public int PassCode { get; set; }
+        public long PassCode { get; set; }
 
         [Required(ErrorMessage = "لطفا {0} را وارد کنید")]
         public int Count { get; set; }
